ライセンス修正. 各種他コンパイラ用バッチ追加

+++ src/strexpr.h
@@ -1,17 +1,21 @@
 /**
- *  @file   strexpr.c
- *  @brief  Ž®•¶Žš—ñ‚ðŒvŽZ‚·‚é
+ *  @file   strexpr.h
+ *  @brief  Ž®•¶Žš—ñ‚ðŒvŽZ
  *  @author Masashi KITAMURA (tenka@6809.net)
  *  @date   1996-2017
+ *  @note
+ *      Boost Software License Version 1.0
  */
 #ifndef STREXPER_H
 #define STREXPER_H
+
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-typedef long long strexpr_val_t;
+typedef int64_t     strexpr_val_t;
 //typedef long strexpr_val_t;
 
 int  StrExpr(char const *s, char const** s_nxt, strexpr_val_t* val);
